Expose retained text splice from GPU renderer
